DOIs included in the package for testing and illustration

DOIs included in the package for testing and illustration

doi_examples

Format

A list of biblids::doi() vectors, including:

good: Guaranteed to work (all are is_metacheckable()).

many: Long list of DOIs.

weird: Troublesome DOIs.

biblids: biblids::doi_examples()

Examples

doi_examples$good
#> <digital object identifier[284]>
#>   [1] 10.1038/s41598-020-57429-5   10.3389/fmech.2019.00073    
#>   [3] 10.1038/s41598-020-62245-y   10.1109/JLT.2019.2961931    
#>   [5] 10.1186/s41476-020-00129-4   10.3390/w12020396           
#>   [7] 10.3390/su12031068           10.3390/su12041435          
#>   [9] 10.36900/suburban.v8i1/2.559 10.3389/fenrg.2020.00005    
#>  [11] 10.3389/fpsyg.2020.00097     10.3389/fmech.2020.00003    
#>  [13] 10.3389/fmech.2020.00004     10.3389/fevo.2020.00018     
#>  [15] 10.1109/ACCESS.2020.2977087  10.16910/jemr.13.1.3        
#>  [17] 10.1088/1748-9326/ab832f     10.1177/1756827720921718    
#>  [19] 10.1038/s41598-020-63481-y   10.1088/2515-7620/ab8917    
#>  [21] 10.1016/j.oceram.2020.100001 10.3390/land9010024         
#>  [23] 10.3390/su12062186           10.3390/su12041534          
#>  [25] 10.3390/su12041623           10.3390/ma13040977          
#>  [27] 10.3390/photonics7010013     10.3390/nu12030762          
#>  [29] 10.3390/lubricants8030029    10.3390/molecules25040934   
#>  [31] 10.3390/ma13071490           10.3390/su12072634          
#>  [33] 10.3390/su12062450           10.3390/su12072791          
#>  [35] 10.3390/s20072033            10.3390/app10072418         
#>  [37] 10.3390/w12030847            10.3390/app10062047         
#>  [39] 10.3389/fmicb.2020.01291     10.1371/journal.pone.0234641
#>  [41] 10.1155/2020/8850198         10.1088/1367-2630/ab9c66    
#>  [43] 10.1038/s41467-020-16977-0   10.5194/wes-5-721-2020      
#>  [45] 10.1186/s12984-020-0640-7    10.1186/s13063-020-04252-y  
#>  [47] 10.3389/fbioe.2020.00209     10.3389/fmats.2020.00157    
#>  [49] 10.3389/frobt.2020.00074     10.3389/fmech.2020.00039    
#>  [51] 10.3389/fmicb.2020.01284     10.1016/j.envint.2020.105867
#>  [53] 10.1016/j.egyr.2020.06.028   10.1186/s12939-020-01191-7  
#>  [55] 10.3390/en13081993           10.3390/su12083394          
#>  [57] 10.3390/su12093507           10.3390/foods9050557        
#>  [59] 10.3390/atmos11050500        10.3390/app10093332         
#>  [61] 10.3390/pr8050582            10.3390/rs12101552          
#>  [63] 10.3390/acoustics2020019     10.3390/su12104145          
#>  [65] 10.3390/nano10050981         10.3390/ijms21082962        
#>  [67] 10.3390/su12104245           10.3390/su12114475          
#>  [69] 10.3390/en13112981           10.3390/membranes10050093   
#>  [71] 10.3390/su12125069           10.3390/mi11060592          
#>  [73] 10.3390/sym12050863          10.3390/admsci10030037      
#>  [75] 10.3390/met10070888          10.3389/fmech.2020.00049    
#>  [77] 10.3389/fmech.2020.00051     10.3389/fmech.2020.00057    
#>  [79] 10.1088/1748-9326/aba2fe     10.5194/hess-24-3835-2020   
#>  [81] 10.3389/fbioe.2020.00854     10.3389/fnins.2020.00795    
#>  [83] 10.3389/fmech.2020.00069     10.3389/fbioe.2020.00944    
#>  [85] 10.5194/hess-24-4239-2020    10.1088/1367-2630/abb5f0    
#>  [87] 10.1088/2515-7620/abb6da     10.3389/fmech.2020.00063    
#>  [89] 10.3389/fpsyg.2020.02182     10.3389/fmech.2020.567386   
#>  [91] 10.3389/fnins.2020.566147    10.1109/ACCESS.2020.3029319 
#>  [93] 10.3390/su12145623           10.3390/e22080802           
#>  [95] 10.3390/s20143990            10.3390/app10155240         
#>  [97] 10.3390/met10081034          10.3390/su12187496          
#>  [99] 10.3390/su12156093           10.3390/su12177135          
#> [101] 10.3390/ma13153418           10.3390/s20154338           
#> [103] 10.3390/math8091450          10.3390/ijms21165890        
#> [105] 10.3390/su12197900           10.3390/molecules25194418   
#> [107] 10.3390/urbansci4030039      10.3390/su12198134          
#> [109] 10.3390/su12187302           10.3390/su12176989          
#> [111] 10.3390/wevj11030056         10.3390/e22090997           
#> [113] 10.3390/e22090984            10.3390/su12187504          
#> [115] 10.3390/pr8091082            10.3390/recycling5030019    
#> [117] 10.3390/mi11090866           10.3390/en13205348          
#> [119] 10.1088/1748-9326/abbd63     10.3390/su12197984          
#> [121] 10.5194/gmd-13-4435-2020     10.36900/suburban.v8i3.523  
#> [123] 10.1002/ejlt.202000213       10.1002/cbic.202000257      
#> [125] 10.1002/adfm.202007294       10.1002/anie.202011696      
#> [127] 10.1002/star.202000080       10.1002/anie.202011337      
#> [129] 10.1002/cite.202000091       10.1002/pssb.202000471      
#> [131] 10.1002/star.202000146       10.1002/star.202000145      
#> [133] 10.1111/ejn.14992            10.1002/anie.202010484      
#> [135] 10.1007/s10841-020-00253-z   10.1007/s00453-020-00746-y  
#> [137] 10.1007/s00187-020-00302-3   10.1007/s10111-020-00626-z  
#> [139] 10.1007/s10664-020-09815-w   10.1007/s10980-020-00974-1  
#> [141] 10.1007/s10626-020-00315-z   10.1007/s13204-020-01328-8  
#> [143] 10.1140/epje/i2020-11975-6   10.1007/s00453-020-00751-1  
#> [145] 10.1038/s41598-020-69807-0   10.1007/s11069-020-04192-5  
#> [147] 10.1007/s11367-020-01737-5   10.1007/s00193-020-00962-z  
#> [149] 10.1007/s12567-020-00308-6   10.1007/s13204-020-01347-5  
#> [151] 10.1007/s11252-020-00983-7   10.1007/s10659-020-09787-4  
#> [153] 10.1007/s00211-020-01138-8   10.3758/s13414-020-02039-6  
#> [155] 10.1007/s11368-020-02727-8   10.1007/s10035-020-01016-y  
#> [157] 10.1007/s00028-020-00578-x   10.1007/s00778-020-00612-x  
#> [159] 10.1007/s12567-020-00310-y   10.1007/s10013-020-00412-4  
#> [161] 10.1007/s42452-020-03253-5   10.1007/s41464-020-00092-0  
#> [163] 10.1007/s12268-020-1368-4    10.1007/s10973-020-09775-9  
#> [165] 10.1007/s12559-020-09733-5   10.1007/s00287-020-01253-8  
#> [167] 10.1140/epje/i2020-11949-8   10.1007/s10853-020-04758-5  
#> [169] 10.1007/s10626-020-00316-y   10.1007/s11042-020-09144-6  
#> [171] 10.1007/s13366-020-00502-x   10.1007/s00397-020-01215-7  
#> [173] 10.1007/s42048-020-00069-1   10.1007/s10013-020-00427-x  
#> [175] 10.1007/s00348-020-02979-7   10.1007/s10460-020-10066-0  
#> [177] 10.1007/s10531-020-02003-8   10.1140/epjst/e2020-900253-0
#> [179] 10.1007/s12274-020-2876-8    10.1007/s00454-020-00226-8  
#> [181] 10.1007/s00211-020-01123-1   10.1007/s13278-020-00664-5  
#> [183] 10.1007/s42452-020-2616-4    10.1007/s10209-020-00748-1  
#> [185] 10.1007/s00396-020-04663-y   10.1007/s11612-020-00522-0  
#> [187] 10.1007/s00453-020-00736-0   10.1007/s11044-020-09722-4  
#> [189] 10.1007/s10198-019-01154-y   10.1007/s00454-020-00173-4  
#> [191] 10.1007/s00426-020-01291-7   10.1007/s11118-020-09830-y  
#> [193] 10.1007/s10683-020-09642-2   10.1007/s10444-020-09763-5  
#> [195] 10.1007/s10198-020-01165-0   10.1007/s11664-020-08433-y  
#> [197] 10.1007/s12011-020-02355-w   10.1007/s41109-020-00311-0  
#> [199] 10.1007/s00021-020-00527-1   10.1038/s41598-020-72559-6  
#> [201] 10.1007/s11016-020-00564-6   10.1007/s00454-020-00248-2  
#> [203] 10.1038/s41598-020-72402-y   10.1007/s00161-020-00910-3  
#> [205] 10.1007/s00550-020-00505-1   10.1007/s11368-020-02762-5  
#> [207] 10.1038/s41467-020-18427-3   10.1007/s13366-020-00523-6  
#> [209] 10.1007/s00397-020-01236-2   10.1007/s00170-020-06074-3  
#> [211] 10.1186/s12913-020-05760-y   10.1631/jzus.A2000033       
#> [213] 10.1007/s41060-020-00229-x   10.1007/s10516-020-09514-7  
#> [215] 10.1007/s12268-020-1450-y    10.1007/s11368-020-02772-3  
#> [217] 10.1007/s00193-020-00973-w   10.1007/s00245-020-09720-0  
#> [219] 10.1007/s11081-020-09559-y   10.1007/s41064-020-00130-z  
#> [221] 10.1007/s10640-020-00521-1   10.1007/s00030-020-00659-3  
#> [223] 10.1007/s11367-020-01837-2   10.1007/s12599-020-00671-y  
#> [225] 10.1007/s00419-020-01808-x   10.1186/s12934-020-01450-w  
#> [227] 10.1007/s40879-020-00436-7   10.1038/s41467-020-19385-6  
#> [229] 10.1007/s00161-020-00939-4   10.1186/s12877-020-01840-0  
#> [231] 10.1007/s00161-020-00941-w   10.1007/s00193-020-00975-8  
#> [233] 10.1039/C8SM02476F           10.1039/C8SM02090F          
#> [235] 10.1039/C8CC09853K           10.1039/C8NR09458F          
#> [237] 10.1039/C8FD00203G           10.1039/C8CP06950F          
#> [239] 10.1039/C8GC03798A           10.1039/C8JA00427G          
#> [241] 10.1039/C9OB00070D           10.1039/C9EE01185D          
#> [243] 10.1039/C9PP00144A           10.1039/C9SM00889F          
#> [245] 10.1039/C9NR03161H           10.1039/C9CP00495E          
#> [247] 10.1039/C9CP02787D           10.1039/C9TA04583J          
#> [249] 10.1039/C9NR02115A           10.1039/C9SM01161G          
#> [251] 10.1039/C9NR06156H           10.1039/C9SM00950G          
#> [253] 10.1039/C9CC06637C           10.1039/C9CP05042F          
#> [255] 10.1039/C9SM01539F           10.1039/C9CC08680C          
#> [257] 10.1039/C9CY01920K           10.1039/C9TA09775A          
#> [259] 10.1039/C9CP06568G           10.1039/C9CY02285F          
#> [261] 10.1039/c8cc09655d           10.1039/c8cs00826d          
#> [263] 10.1039/c8ee03682a           10.1039/c9cc09092d          
#> [265] 10.1039/c9cp02996f           10.1039/c9ew00596j          
#> [267] 10.1039/c9fd00059c           10.1039/C9CY02579K          
#> [269] 10.1039/D0CP02110E           10.1039/D0EE01125H          
#> [271] 10.1039/D0NR03170D           10.1039/D0SM00616E          
#> [273] 10.1039/D0EE01013H           10.1039/D0CC03229H          
#> [275] 10.1039/D0CP02919J           10.1039/D0JA00208A          
#> [277] 10.1039/D0CS00458H           10.1039/D0CP03650A          
#> [279] 10.1039/D0CS00815J           10.1039/D0SM01238F          
#> [281] 10.1039/D0GC02779K           10.1039/D0EE01912G          
#> [283] 10.1039/D0CP03233F           10.1039/d0ra05997h