Send metacheck results as an parametrised email

Send metacheck results as an parametrised email

mc_compose_email(dois, translator = mc_translator(), ...)

mc_compose_email_outer(body = "Lorem", translator = mc_translator())

mc_render_email(
  dois = doi_examples$good[1:10],
  translator = mc_translator(),
  ...
)

render_and_send(to, translator = mc_translator(), ...)

render_and_send_async(...)

smtp_send_mc(
  email = blastula::prepare_test_message(),
  to = throwaway,
  from = "metacheck-support@sub.uni-goettingen.de",
  credentials = creds_metacheck(),
  translator = mc_translator(),
  ...
)

Arguments

dois
Vector of DOIs, as created by, or coerceable to biblids::doi().

translator
A shiny.i18n::Translator object or NULL for english-only defaults. Strings inside the module UI are marked as translateable. You can pass a translator object included in the package, or can create your own translator using shiny.i18n::Translator. This must not be a reactive, it is only set at shiny startup. To update the language reactively during a shiny session, see lang.

...
Arguments passed on to mc_render_email, blastula::render_email, mc_compose_email, blastula::smtp_send

input

The input file to be rendered. This should be an R Markdown document (.Rmd) with the output specified as output: blastula::blastula_email.

envir

The environment in which the code chunks are to be evaluated during knitting.

quiet

An option to suppress printing of the command line output from Pandoc during rendering. By default, this is set to TRUE.

output_options

Lists of options can be used to augment the rendering of the email message. The output_options list will be passed as the output_options argument of rmarkdown::render(). The render_options list is for providing additional arguments to rmarkdown::render(). By default, both lists are empty.

render_options

Lists of options can be used to augment the rendering of the email message. The output_options list will be passed as the output_options argument of rmarkdown::render(). The render_options list is for providing additional arguments to rmarkdown::render(). By default, both lists are empty.

subject

The subject of the message, which is usually a brief summary of the topic of the message. If not provided, an empty string will be used (which is handled differently by email clients).

cc

A vector of email addresses for sending the message as a carbon copy or blind carbon copy. The CC list pertains to recipients that are to receive a copy of a message that is addressed primarily to others. The CC listing of recipients is visible to all other recipients of the message. The BCC list differs in that those recipients will be concealed from all other recipients (including those on the BCC list). A named character vector can be used to specify the recipient names along with the their email address (e.g., c("Joe Public" = "joe_public@example.com")).

bcc

A vector of email addresses for sending the message as a carbon copy or blind carbon copy. The CC list pertains to recipients that are to receive a copy of a message that is addressed primarily to others. The CC listing of recipients is visible to all other recipients of the message. The BCC list differs in that those recipients will be concealed from all other recipients (including those on the BCC list). A named character vector can be used to specify the recipient names along with the their email address (e.g., c("Joe Public" = "joe_public@example.com")).

creds_file

An option to specify a credentials file. As this argument is deprecated, please consider using credentials = creds_file(<file>) instead.

verbose

Should verbose output from the internal curl send_mail() call be printed? While the username and password will likely be echoed during the exchange, such information is encoded and won't be stored on the user's system.

body
The three layout sections for an email message (ordered from top to bottom). Markdown text can be supplied to each of these by using the md() text helper function. Alternatively, we can supply a set of block_*() calls enclosed within the blocks() function to take advantage of precomposed HTML blocks.

to
A vector of email addresses serving as primary recipients for the message. For secondary recipients, use the cc and bcc arguments. A named character vector can be used to specify the recipient names along with the their email address (e.g., c("Jane Doe" = "jane_doe@example.com")).

email
The email message object, as created by the compose_email() function. The object's class is email_message.

from
The email address of the sender. Often this needs to be the same email address that is associated with the account actually sending the message. As with to, cc, and bcc, we can either supply a single email address or use a named character vector with the sender name and email address (e.g., c("John Doe" = "john_doe@example.com")).

credentials
One of three credential helper functions must be used here: (1) creds(), (2) creds_key(), or (3) creds_file(). The first, creds(), allows for a manual specification of SMTP configuration and credentials within that helper function. This is the most secure method for supplying credentials as they aren't written to disk. The creds_key() function is used if credentials are stored in the system-wide key-value store, through use of the create_smtp_creds_key() function. The creds_file() helper function relies on a credentials file stored on disk. Such a file is created using the create_smtp_creds_file() function.

Related Functions and Methods

Functions

mc_compose_email: Compose complete mail

mc_compose_email_outer: Wrap inner email in outer content

mc_render_email: Render email body (inner content)

render_and_send: Render and send

render_and_send_async: Render and send asynchronously

smtp_send_mc: Send

See also