Email Report through a Shiny Module

Email Report through a Shiny Module

emailReport()

emailReportUI(id, width = "100%", ...)

emailReportServer(id, dois)

Arguments

id
The id string to be namespaced (optional).

width
The width of the input, e.g. '400px', or '100%'; see validateCssUnit().

...
Arguments passed on to shiny::actionButton

inputId

The input slot that will be used to access the value.

label

The contents of the button or link--usually a text label, but you could also use any other HTML, like an image.

icon

An optional icon() to appear on the button.

Related Functions and Methods

Functions

emailReport: Test app

emailReportUI: Module UI

emailReportServer: Module server

See also