Set or get metacheck LLM use

Mainly for use in optional LLM workflows in modules, also checks if the GROQ API key is set and returns false if it isn't.

Usage

llm_use(llm_use = NULL)

Arguments

llm_use

if logical, sets whether to use LLMs

Value

the current option value (logical)

Examples

if (llm_use()) {
  print("We can use LLMs")
} else {
  print("We will not use LLMs")
}
#> [1] "We will not use LLMs"