metacheck: Crossref Metadata Compliance

Automatically check metadata compliance for hybrid open access (OA). Includes a webapp to send compliance reports by email.

See also

Author

Maintainer: Maximilian Held info@maxheld.de (ORCID)

Authors:

Najko Jahn najko.jahn@gmail.com (ORCID)

Henrieke Walter henrieke.walter@sub.uni-goettingen.de (ORCID)

Other contributors:

Göttingen State and University Library [copyright holder, funder]

German Research Foundation (Projektnummer #416115939) [sponsor]

Crossref [data contributor]