Pretest acceptable DOIs

Pretest acceptable DOIs

tabulate_metacheckable(x, ...)

is_metacheckable(x, ...)

assert_metacheckable(x, ...)

report_metacheckable(x, ...)

is_in_limit(x, limit = 1000L)

is_doi_cr_md(x)

is_doi_cr_type(x, type = types_allowed)

Arguments

x
A vector created by, or convertable to doi().

...
Additional arguments passed to the predicate functions.

limit
Scalar integer, giving the number of DOIs to be submitted to the metacheck.

type
Character scalar, must be one of the types from rcrossref::cr_types().

Value

A tibble with one column for each of the predicate results. TRUE always means that a DOI passes a criterion.

Tested in order:

not_na following biblids::doi().

unique FALSE for every 2nd and later repetition of a DOI. DOIs are compared using biblids::doi() logic.

within_limits whether remaining DOIs are within the package limit for requests.

doi_org_found whether remaining DOIs are found on DOI.org, using biblids::is_doi_found().

from_cr whether remaining DOIs have been deposited by the Crossref registration agency (per doi.org).

cr_md whether remaining DOIs have metadata on Crossref.

NA can indicate that the test:

was not applicable, because a previous predicate was FALSE

Details

Before even checking what metadata has been deposited for a DOI, a DOI has to pass a number of criteria to be eligible for metacheck.

The below criteria are assumed to be transitive but asymmetric in the order given. For example, if a DOI is not found on DOI.org, it is assumed that it also cannot be known to crossref and all later predicates are also assumed to be NA. This must necessarily be the case for, say, syntactically invalid DOIs, but is more of an assumption when it comes to the consistency between different data sources (say, DOI.org and Crossref).

Related Functions and Methods

Functions

tabulate_metacheckable: Tabulate results

is_metacheckable: Are all predicates TRUE?

assert_metacheckable: Assert metacheckability

report_metacheckable: Report summary in markdown

is_in_limit: Is the DOI in the limit?

is_doi_cr_md: Is there metadata for the DOI on Crossref?

is_doi_cr_type: Is the DOI a type (per Crossref)?

See also