Display a Table in a Report

A function to display tables in reports.

Usage

report_table(table, colwidths = "auto", maxrows = 2, escape = FALSE)

Arguments

table

the data frame to show in a table, or a vector for a list

colwidths

set column widths as a vector of px (number > 1) or percent (numbers <= 1)

maxrows

if the table has more rows than this, paginate

escape

whether or not to escape the DT (necessary if using raw html)

Value

the datatable

Examples

report_table(iris)