Email Report through a Shiny Module

Email Report through a Shiny Module

emailReport()

emailReportUI(id, width = "100%", translator = mc_translator(), ...)

emailReportServer(
  id,
  dois = shiny::reactive(NULL),
  translator = mc_translator(),
  lang = shiny::reactive("en")
)

Arguments

id
The id string to be namespaced (optional).

width
The width of the input, e.g. '400px', or '100%'; see validateCssUnit().

translator
A shiny.i18n::Translator object or NULL for english-only defaults. Strings inside the module UI are marked as translateable. You can pass a translator object included in the package, or can create your own translator using shiny.i18n::Translator. This must not be a reactive, it is only set at shiny startup. To update the language reactively during a shiny session, see lang.

...
Arguments passed on to shiny::actionButton

inputId

The input slot that will be used to access the value.

label

The contents of the button or link--usually a text label, but you could also use any other HTML, like an image.

icon

An optional icon() to appear on the button.

lang
a reactive variable, returning a character scalar. Must be one of the languages in translator. Defaults to shiny::reactive("en"), in which case no server side translation is triggered.

Related Functions and Methods

Functions

emailReport: Test app

emailReportUI: Module UI

emailReportServer: Module server

See also